Cleaning up journal api.
Not sure about generic JournalAwaiter. It should help in cases when there is one result (or more with deconstruction in C# 7) but then compiler uses generic JournalAwaiter for single assignment lambdas and user has to specify return values for all Whens even when he is not interested in return values, or he has to use code block.