Don't swallow client gump response cancellation after close generic gump packet.
Slow clients can send gump response to BF/04 after receiving a new gump from the server. The scenario is:
10:14:39.802 >>>> server -> proxy: RawPacket SendGumpMenuDialog
10:14:39.805 >>>> proxy -> client: RawPacket SendGumpMenuDialog
10:14:40.366 >>>> proxy -> server: RawPacket GumpMenuSelection
10:14:40.367 >>>> proxy -> client: RawPacket GeneralInformationPacket
10:14:40.406 >>>> server -> proxy: RawPacket SendGumpMenuDialog
10:14:40.408 >>>> proxy -> client: RawPacket SendGumpMenuDialog
10:14:40.416 >>>> client -> proxy: RawPacket GumpMenuSelection
10:14:40.416 >>>> proxy -> server: RawPacket GumpMenuSelection

Hopefully it doesn't break anything else. Current assumption is that if client receives close generic gump packet (BF/04), then the next gump response from client can be blocked. This assumption is not completely correct because client can trigger another gump in the meantime, so other gump response is blocked. Additional assumption is that this case can be triggered just by user messing with automatically requested gumps or other automation tool like Injection. I consider it as corner cases and it is not important to fully support them.