Reverting back last changes.
Cannot make links work the same way on wiki and in the source code browser. Link to an md file on wiki has to be without `.md` extension, otherwise it is rendered as a plain text file. The `.md` extension has to be part of the link url, otherwise code browser cannot find the file.